Update README.md
Flipbook README
This is a simple flipbook project created using HTML and CSS. It consists of several pages that can be flipped through using checkboxes and CSS animations. Below, you'll find an overview of the project structure and how to use it.

Project Structure
The project structure includes the following files:

index.html: This is the main HTML file that defines the structure of the flipbook, including the cover, pages, and back cover.

style.css: This is the CSS file responsible for styling the flipbook. It includes styles for the cover, pages, animations, and more.

Images: The project includes various image files (e.g., pic2.0.jpg, pic12.0.jpg, cover pic.jpg) used as content within the flipbook.

How to Use
To view the flipbook, follow these steps:

Clone the repository to your local machine.

Open the index.html file in a web browser of your choice.

You will see the cover of the flipbook. To flip through the pages, click on the left or right corners of the pages. This will simulate turning the pages of a physical book.

The flipbook includes several pages, each with different content and images.

Enjoy flipping through the pages and reading the content!

Customization
If you want to customize the flipbook, you can do so by modifying the HTML and CSS files. Here are some customization options:

Changing Content: You can replace the content of each page with your own text and images by modifying the HTML within the <div class="page"> elements.

Styling: Feel free to modify the styles in the style.css file to change the colors, fonts, or animations used in the flipbook.

Adding More Pages: If you want to add more pages to the flipbook, you can replicate the structure of existing pages by creating additional <div class="page"> elements in the HTML.

Credits
Icons: This project uses icons from FontAwesome.

Quote: A quote by William Shakespeare is included in one of the pages.

Author
This flipbook project was created by Aishwarya MS.